requirements.txt
# NeoXBridge AI - Python Dependencies
# Install with: pip install -r requirements.txt

# Core AI and Language Model
openai>=1.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.0

# JSON handling
orjson>=3.9.0

# Async support
aiofiles>=23.0.0

# Logging enhancements
rich>=13.0.0

# Date/time utilities
python-dateutil>=2.8.0

# Windows color support
colorama>=0.4.0
